# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vyapaar AI API Specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List User Organizations
      description: Retrieves all organizations the authenticated user is a member of.
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create Organization
      description: Creates a new business organization tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        "200":
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}/switch:
    post:
      operationId: switchOrganization
      tags: [organizations]
      summary: Switch Active Organization
      description: Selects active organization profile.
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active organization switched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orgId:
                    type: integer
...
  /business-profile:
    get:
      operationId: getBusinessProfile
      tags: [organizations]
      summary: Fetch Business Profile
      description: Retrieves active organization business settings and branding details.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business profile details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    put:
      operationId: updateBusinessProfile
      tags: [organizations]
      summary: Update Business Profile
      description: Updates branding configurations, address, and business hours.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProfileInput"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
...
  /knowledge/upload:
    post:
      operationId: uploadKnowledgeFile
      tags: [knowledge]
      summary: Upload Knowledge File
      description: Uploads a PDF or text catalog file, generates embeddings, and indexes chunks.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Document successfully indexed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  documentId:
                    type: integer
                required: [success, message, documentId]
  /knowledge/query:
    post:
      operationId: queryKnowledge
      tags: [knowledge]
      summary: Query Knowledge Base
      description: Query the scoped RAG knowledge base using a natural-language question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                orgId:
                  type: integer
              required: [question, orgId]
      responses:
        "200":
          description: grounded support assistant response
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        filename:
                          type: string
                        contentText:
                          type: string
                        similarity:
                          type: number
                      required: [id, filename, contentText, similarity]
                required: [answer, sources]
  /knowledge/documents:
    get:
      operationId: listKnowledgeDocuments
      tags: [knowledge]
      summary: List Knowledge Documents
      description: Retrieves the catalog files list.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    orgId:
                      type: integer
                    filename:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
                  required: [id, orgId, filename]
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name]
...
    BusinessProfile:
      type: object
      properties:
        orgId:
          type: integer
        logoUrl:
          type: string
        tagline:
          type: string
        primaryColor:
          type: string
        shortDescription:
          type: string
        category:
          type: string
        phone:
          type: string
        address:
          type: string
        businessHours:
          type: object
      required: [orgId]
    BusinessProfileInput:
      type: object
      properties:
        orgId:
          type: integer
        logoUrl:
          type: string
        tagline:
          type: string
        primaryColor:
          type: string
        shortDescription:
          type: string
        category:
          type: string
        phone:
          type: string
        address:
          type: string
        businessHours:
          type: object
      required: [orgId]
...